added collision and small changes

+++ src/map.h
@@ -41,9 +41,9 @@ struct map
 struct map *map_create(int chunksize, int tilesize, int width, int height);
 struct map *map_load(char *dir);
 void map_destroy(struct map *map);
-struct chunk *map_get_chunk_from_coordinate(struct map *map, int x, int y);
+struct chunk *map_get_chunk_from_coordinate(struct map *map, float x, float y);
 struct chunk *map_get_chunk_from_index(struct map *map, int x, int y);
-struct tile *map_get_tile_from_coordinate(struct map *map, int x, int y);
+struct tile *map_get_tile_from_coordinate(struct map *map, float x, float y);
 int map_save(struct map *map, char *dir);
 void map_destroy_chunk(struct chunk *chunk, int chunksize);
